refactor eclCluster module, fix resulting compile errors
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@19605 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ b2bii/modules/B2BIIConvertMdst/src/B2BIIConvertMdstModule.cc
@@ -958,7 +958,7 @@ void B2BIIConvertMdstModule::convertMdstECLObject(const Belle::Mdst_ecl& ecl, co
 
   // TODO: check
   // TODO: ECLCluster is an unclear mess
-  float covarianceMatrix[6];
+  double covarianceMatrix[6];
   covarianceMatrix[0] = sqrt(ecl.error(0)); // error on energy
   covarianceMatrix[1] = ecl.error(1);
   covarianceMatrix[2] = sqrt(ecl.error(2)); // error on phi
